backend/upload_file/requirements.txt
fastapi
gunicorn
pydantic-sqlalchemy
python-dotenv
python-multipart
uvicorn[standard]
asyncpg
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
PyJWT
vakt
pillow
opencv-python
requests